requirements.txt
pbr>=0.6,<1.0
SQLAlchemy>=0.7.8,!=0.9.5,<=0.9.99
amqplib>=0.6.1
anyjson>=0.3.3,<=0.3.3
argparse
boto>=2.12.0,!=2.13.0,<2.35.0
eventlet>=0.13.0,<0.16.0
Jinja2<=2.7.3
kombu>=2.5.0,<=3.0.7
lxml>=2.3,<=3.4.2
Routes>=1.12.3,!=2.0
WebOb>=1.2.3,<=1.4
greenlet>=0.3.2,<=0.4.5
PasteDeploy>=1.5.0,<=1.5.2
Paste
sqlalchemy-migrate>=0.8.2,!=0.8.4,<=0.9.1
netaddr>=0.7.6,<=0.7.14
suds==0.4
paramiko>=1.9.0,<=1.15.2
pyasn1<=0.1.7
Babel>=1.3,<=1.3
iso8601>=0.1.9,<=0.1.10
jsonschema>=2.0.0,<3.0.0
python-cinderclient>=1.0.6,<=1.1.1
python-neutronclient>=2.3.4,<2.3.11
python-glanceclient>=0.9.0,!=0.14.0,<=0.14.2
python-keystoneclient>=0.7.0,<0.12.0
six>=1.6.0,<=1.9.0
stevedore>=0.14
websockify>=0.5.1,<0.6
wsgiref>=0.1.2
oslo.config>=1.2.0,<1.5
oslo.rootwrap<1.4
pycadf>=0.4.1,<0.6.0
oslo.messaging>=1.3.0,<1.5
ncclient>=0.4.2
